Update index.js
Changing datatype for number to translate to decimal instead of float. This is done for avro schema compatibility.

+++ src/index.js
@@ -10,7 +10,7 @@ const typeMapping = {
 	'null': 'null',
 	'boolean': 'boolean',
 	'integer': 'int',
-	'number': 'float'
+	'number': 'decimal'
 }
 
 const reSymbol = /^[A-Za-z_][A-Za-z0-9_]*$/;
